examples/qwen2/requirements.txt
datasets~=2.3.2
rouge_score~=0.1.2
# transformers~=4.31.0  # tensorrt-llm has installed
transformers-stream-generator
sentencepiece~=0.1.99
tiktoken
einops

# optional dependencies
uvicorn
gradio==3.40.1
mdtex2html
sse_starlette
aiohttp_sse_client
openai==1.1.1
